Add support for RT kernel
The `gckOS_WaitSignal` API in the driver is same as
`wait_for_completion_interruptible_timeout` kernel completion API except for
the additional `manualReset`. In the case of RT kernel, `wait_for_completion`
API is updated to use raw spinlocks and swait instead of spinlock and
waitqueue. Update the `gckOS_WaitSignal` to use correct completion mechanism
for RT kernels.

The presence of RT kernel is determined by the presence of the
`CONFIG_PREEMPT_RT_BASE` kernel config option.

Signed-off-by: Gowtham Tammana <g-tammana@ti.com>